requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-jwt-extended>=4.6.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.30.0
requests>=2.31.0
gunicorn>=22.0.0